make hello world match what we send people
