+++ app/src/androidTest/java/com/example/jibo/helloworld/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.calvin.romtest;
+package com.example.jibo.helloworld;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -10,7 +10,7 @@ import org.junit.runner.RunWith;
 import static org.junit.Assert.*;
 
 /**
- * Instrumentation test, which will execute on an Android device.
+ * Instrumented test, which will execute on an Android device.
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.calvin.romtest", appContext.getPackageName());
+        assertEquals("com.example.jibo.helloworld", appContext.getPackageName());
     }
 }
