+++ app/src/test/java/com/example/jibo/helloworld/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.calvin.romtest;
+package com.example.jibo.helloworld;
 
 import org.junit.Test;
